+++ changelog.md
@@ -41,6 +41,7 @@
 - [#3109](https://github.com/ignite/cli/pull/3109) Refactor scaffolding for proto files to not rely on placeholders.
 - [#3106](https://github.com/ignite/cli/pull/3106) Add zoom image plugin.
 - [#3224](https://github.com/ignite/cli/pull/3224) Remove grpc_* prefix from query files in scaffolded chains
+- [#3229](https://github.com/ignite/cli/pull/3229) Rename `campaign` to `project` in ignite network set of commands
 - [#3244](https://github.com/ignite/cli/pull/3244) updated actions.yml for resolving deprecation message
 
 ### Breaking Changes